Crate fj

Fornjot Modeling Library

This library is part of the Fornjot ecosystem. Fornjot is an open-source, code-first CAD application; and collection of libraries that make up the CAD application, but can be used independently.

The purpose of this library is to support Fornjot models, which are just Rust libraries. Models depend on this library and use the primitives defined here to define a CAD model. Together with the Fornjot application, this library forms the part of Fornjot that is relevant to end users.

To display the created CAD model, or export it to another file format, you need the Fornjot application. Please refer to the Fornjot repository for usage examples.

Modules

syntax

Convenient syntax for fj operations

Structs

Angle

An angle

Circle

A circle that is part of a Sketch

Difference2d

A difference between two shapes

Group

A group of two 3-dimensional shapes

PolyChain

A polygonal chain that is part of a Sketch

Sketch

A sketch

Sweep

A sweep of a 2-dimensional shape along straight path

Transform

A transformed 3-dimensional shape

Enums

Chain

A chain of elements that is part of a Sketch

Shape

A shape

Shape2d

A 2-dimensional shape

Attribute Macros

model

Define a function-based model.