fj

Struct Instance

pub struct Instance {
    pub core: Core,
}

An instance of Fornjot

This is the main entry point into the Fornjot API

Fields
core: Core

The instance of the Fornjot core

Implementations

impl Instance

pub fn new() -> Self

Construct an instance of Instance

pub fn with_validation_config(config: ValidationConfig) -> Self

Construct an instance of Instance, using the provided configuration

pub fn process_model<M>(&mut self, model: &M) -> Result
where for<'r> (&'r M, Tolerance): Triangulate, for<'r> &'r M: BoundingVolume<3>,

Export or display a model, according to CLI arguments

This function is intended to be called by applications that define a model and want to provide a standardized CLI interface for dealing with that model.

This function is used by Fornjot’s own testing infrastructure, but is useful beyond that, when using Fornjot directly to define a model.

pub fn process_model_args<M>(&mut self, model: &M, args: Args) -> Result
where for<'r> (&'r M, Tolerance): Triangulate, for<'r> &'r M: BoundingVolume<3>,

Process a model with pre-parsed arguments

This function is similar to Self::process_model, but accepts pre-parsed arguments instead of parsing them from the command line. This is useful when you want to extend the standard arguments with your own parameters.